chore: remove legacy terminal worker
